add api for direct customer delete and fatch existing customer

+++ app/clients/hubspot_api.py
@@ -1,7 +1,6 @@
 # app/clients/hubspot_api.py
 import httpx
 from app.config import HUBSPOT_BASE_URL, HUBSPOT_TOKEN
-
 HEADERS = {
     "Authorization": f"Bearer {HUBSPOT_TOKEN}",
     "Content-Type": "application/json"
@@ -9,11 +8,15 @@ HEADERS = {
 
 async def create_contact(payload: dict) -> str:
     async with httpx.AsyncClient(timeout=15) as client:
-        res = await client.post(
-            f"{HUBSPOT_BASE_URL}/crm/v3/objects/contacts",
-            headers=HEADERS,
-            json=payload
-        )
+        try:
+            res = await client.post(
+                    f"{HUBSPOT_BASE_URL}/crm/v3/objects/contacts",
+                    headers=HEADERS,
+                    json=payload
+            )
+        except httpx.HTTPError as e:
+            print("Error creating contact:", e)
+            raise e
         res.raise_for_status()
         return res.json()["id"]
 
@@ -84,16 +87,30 @@ async def update_ticket(hubspot_ticket_id: str, properties: dict):
         return res.json()
     
 async def delete_ticket(hubspot_ticket_id: str):
-    """
-    Delete a ticket from HubSpot
-    """
+
+    hubspot_ticket_id = str(hubspot_ticket_id).strip()
+
+    if not await ticket_exists(hubspot_ticket_id):
+        print("⚠️ Ticket not found, skipping delete")
+        return False
+
     url = f"{HUBSPOT_BASE_URL}/crm/v3/objects/tickets/{hubspot_ticket_id}"
 
     async with httpx.AsyncClient(timeout=15) as client:
         res = await client.delete(url, headers=HEADERS)
 
         if res.status_code >= 400:
-            print("HUBSPOT DELETE ERROR:", res.status_code, res.text)
+            print("❌ HUBSPOT DELETE ERROR:", res.status_code, res.text)
 
         res.raise_for_status()
         return True
+
+    
+async def ticket_exists(ticket_id: str):
+
+    url = f"{HUBSPOT_BASE_URL}/crm/v3/objects/tickets/{ticket_id}"
+
+    async with httpx.AsyncClient(timeout=10) as client:
+        res = await client.get(url, headers=HEADERS)
+
+        return res.status_code == 200
